backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.2
cffi==1.17.1
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.13
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
passlib==1.7.4
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
